# swagger.yaml
openapi: 3.0.0
info:
  title: API Простого Інтернет-Магазину
  version: 1.0.0
  description: Це просте API для інтернет-магазину, яке дозволяє користувачам переглядати товари та робити замовлення.

servers:
  - url: 'http://localhost:5000/api'
    description: Сервер розробки

paths:
  /products:
    get:
      summary: Перелік усіх товарів
      operationId: listProducts
      responses:
        '200':
          description: Масив товарів
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductArray'
        '400':
          description: Масив товірів не зайдено
    post:
      summary: Створення товару
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '201':
          description: Товар створено
        '404':
          description: Товар не створено


  /products/{productId}:
    get:
      summary: Отримати товар за його ідентифікатором
      operationId: getProductById
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Об'єкт товару
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Товар не знайдено
    put:
      summary: Оновлення товару
      operationId: updateProduct
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Товар оновлено
        '404':
          description: Товар не знайдено
    delete:
      summary: Видалення товару
      operationId: deleteProduct
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Товар видалено
        '404':
          description: Товар не знайдено

  /orders:
    post:
      summary: Створення замовлення
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '201':
          description: Замовлення створено
        '400':
          description: Недійсне замовлення

components:
  schemas:
    Product:
      type: object
      required:
        - id
        - name
        - price
        - stock
      properties:
        id:
          type: string
          description: Унікальний ідентифікатор товару
        name:
          type: string
          description: Назва товару
        price:
          type: number
          description: Ціна товару
        stock:
          type: integer
          description: Кількість товару на складі
    ProductArray:
      type: array
      items:
        $ref: '#/components/schemas/Product'
    Order:
      type: object
      required:
        - productId
        - quantity
      properties:
        productId:
          type: string
          description: Унікальний ідентифікатор товару
        quantity:
          type: integer
          description: Кількість товару в замовленні